# app/docs/openapi.json
{
  "openapi" : "3.1.0",
  "paths" : {
...
    "/internal/accounts/{userId}/download/{nodeId}" : {
      "get" : {
        "parameters" : [ {
          "name" : "nodeId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "summary" : "Download",
        "tags" : [ "Internal Blob Resource" ]
      }
    },
...
    "/internal/accounts/{userId}/upload" : {
      "post" : {
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Content-Length",
          "in" : "header",
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "Description",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "Filename",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "ParentId",
          "in" : "header",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "*/*" : {
              "schema" : {
                "type" : "string",
                "format" : "binary"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : { }
            }
          }
        },
        "summary" : "Upload",
        "tags" : [ "Internal Blob Resource" ]
      }
    },
...
  },
...
  "info" : {
    "title" : "carbonio-files-ce-app API",
    "version" : "unversioned"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Auto generated value"
  }, {
    "url" : "http://0.0.0.0:8080",
    "description" : "Auto generated value"
  } ]
}